Moved getActiveEnroResultChannels extension into TestExtensions.kt
